Revert(android): 修改包名并移动MainActivity文件
将应用包名从com.oierxjn.onetj改回为com.example.onetj
将MainActivity文件移动到新的包目录下
